refactor: rename folders Player by GameDatas
