+++ gui/Makefile
@@ -17,14 +17,14 @@ SRC			=	src/Error/Error.cpp \
 				src/Parsing/ServerParser.cpp \
 				src/Parsing/ParseCommandLine.cpp \
 				\
-				src/Player/Inventory.cpp \
-				src/Player/Player.cpp \
+				src/GameDatas/Inventory.cpp \
+				src/GameDatas/Player.cpp \
 
 TEST_FILES	=	Network/TestNetwork.cpp \
 				Parsing/TestParser.cpp \
 				Parsing/TestParseCommandLine.cpp \
-				Player/TestInventory.cpp \
-				Player/TestPlayer.cpp \
+				GameDatas/TestInventory.cpp \
+				GameDatas/TestPlayer.cpp \
 
 OBJ			=	$(SRC:.cpp=.o)
 MAIN_OBJ	=	$(MAIN:.cpp=.o)
